Make ScheduleActivity abstract, move tabs code to WeekScheduleActivity
