+++ app/src/main/java/by/toggi/rxbsuir/activity/ScheduleActivity.java
@@ -6,16 +6,15 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.LayoutRes;
 import android.support.annotation.Nullable;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.NavigationView;
 import android.support.design.widget.Snackbar;
-import android.support.design.widget.TabLayout;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.view.GravityCompat;
-import android.support.v4.view.ViewPager;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -37,7 +36,6 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import by.toggi.rxbsuir.R;
 import by.toggi.rxbsuir.RxBsuirApplication;
-import by.toggi.rxbsuir.Utils;
 import by.toggi.rxbsuir.adapter.WeekPagerAdapter;
 import by.toggi.rxbsuir.component.DaggerScheduleActivityComponent;
 import by.toggi.rxbsuir.fragment.AddEmployeeDialogFragment;
@@ -57,7 +55,7 @@ import icepick.State;
 import static by.toggi.rxbsuir.mvp.presenter.SchedulePresenter.Error;
 
 
-public class ScheduleActivity extends AppCompatActivity implements ScheduleView, NavigationDrawerView, NavigationView.OnNavigationItemSelectedListener, OnButtonClickListener {
+public abstract class ScheduleActivity extends AppCompatActivity implements ScheduleView, NavigationDrawerView, NavigationView.OnNavigationItemSelectedListener, OnButtonClickListener {
 
     public static final String KEY_IS_GROUP_SCHEDULE = "is_group_schedule";
     public static final String KEY_SUBGROUP_1 = "subgroup_1";
@@ -69,8 +67,6 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
     private static final String KEY_TITLE = "title";
 
     @Bind(R.id.toolbar) Toolbar mToolbar;
-    @Bind(R.id.tab_layout) TabLayout mTabLayout;
-    @Bind(R.id.view_pager) ViewPager mViewPager;
     @Bind(R.id.progress_bar) ProgressBar mProgressBar;
     @Bind(R.id.coordinator_layout) CoordinatorLayout mCoordinatorLayout;
     @Bind(R.id.fab_group) FloatingActionButton mFabGroup;
@@ -100,7 +96,7 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
         setTheme(mIsDarkTheme ? R.style.AppTheme_Drawer_Dark : R.style.AppTheme_Drawer_Light);
 
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_schedule);
+        setContentView(getLayoutRes());
 
         ButterKnife.bind(this);
         mFloatingActionMenu.getBackground().setAlpha(0);
@@ -109,20 +105,19 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
 
         setupNavigationView();
 
-        setupTabs();
-
         mSchedulePresenter.attachView(this);
         mSchedulePresenter.onCreate();
 
         mDrawerPresenter.attachView(this);
         mDrawerPresenter.onCreate();
 
-        showCurrentWeek();
-
         Icepick.restoreInstanceState(this, savedInstanceState);
         setupTitle();
 
     }
+
+    @LayoutRes
+    protected abstract int getLayoutRes();
 
     @Override
     protected void onDestroy() {
@@ -162,8 +157,6 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
 
     @Override
     public void showError(Error error) {
-//        disableScrollFlags();
-        mViewPager.setVisibility(View.GONE);
         mProgressBar.setVisibility(View.GONE);
         switch (error) {
             case NETWORK:
@@ -184,16 +177,12 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
 
     @Override
     public void showLoading() {
-//        disableScrollFlags();
-        mViewPager.setVisibility(View.GONE);
         mProgressBar.setVisibility(View.VISIBLE);
     }
 
     @Override
     public void showContent(int position) {
-//        enableScrollFlags();
         mProgressBar.setVisibility(View.GONE);
-        mViewPager.setVisibility(View.VISIBLE);
     }
 
     @Override
@@ -234,7 +223,7 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
                 mSchedulePresenter.retry();
                 return true;
             case R.id.action_today:
-                showCurrentWeek();
+                showToday();
                 return true;
             case R.id.action_subgroup_1:
                 item.setChecked(!item.isChecked());
@@ -248,6 +237,8 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
                 return super.onOptionsItemSelected(item);
         }
     }
+
+    protected abstract void showToday();
 
     @Override
     public boolean onPrepareOptionsMenu(Menu menu) {
@@ -358,13 +349,6 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
         mNavigationView.setNavigationItemSelectedListener(this);
     }
 
-    private void setupTabs() {
-        mViewPager.setAdapter(mPagerAdapter);
-        mViewPager.setOffscreenPageLimit(4);
-        mViewPager.setPageMargin(mPageMargin);
-        mTabLayout.setupWithViewPager(mViewPager);
-    }
-
     private boolean isMenuItemEnabled() {
         return mSyncId != null;
     }
@@ -379,10 +363,6 @@ public class ScheduleActivity extends AppCompatActivity implements ScheduleView,
         AppBarLayout.LayoutParams params = (AppBarLayout.LayoutParams) mToolbar.getLayoutParams();
         params.setScrollFlags(AppBarLayout.LayoutParams.SCROLL_FLAG_SCROLL | AppBarLayout.LayoutParams.SCROLL_FLAG_ENTER_ALWAYS);
         mToolbar.setLayoutParams(params);
-    }
-
-    private void showCurrentWeek() {
-        mViewPager.setCurrentItem(Utils.getCurrentWeekNumber() - 1);
     }
 
     private void addStorageFragment() {
