+++ app/src/main/java/by/toggi/rxbsuir/activity/StartupActivity.java
@@ -4,15 +4,12 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import timber.log.Timber;
-
 public class StartupActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Timber.d("onCreate()");
-        startActivity(new Intent(this, ScheduleActivity.class));
+        startActivity(new Intent(this, WeekScheduleActivity.class));
         finish();
     }
 }
